added lockfile module; created all-in-one function
Removed potential panic from the `Drop` implementation.
