+++ crates/typst-kit/src/package.rs
@@ -1,12 +1,10 @@
 //! Download and unpack packages and package indices.
 
 use std::fs;
-use std::io;
 use std::path::{Path, PathBuf};
 use std::{thread::sleep, time::Duration};
 
 use ecow::eco_format;
-use fs4::fs_std::FileExt;
 use once_cell::sync::OnceCell;
 use typst::diag::{bail, PackageError, PackageResult, StrResult};
 use typst::syntax::package::{
@@ -14,111 +12,7 @@ use typst::syntax::package::{
 };
 
 use crate::download::{Downloader, Progress};
-
-struct LockFile {
-    path: PathBuf,
-    tmp_path: String,
-    file: fs::File,
-}
-
-impl LockFile {
-    /// `packages_root_dir` is normally either the cache or the local directory.
-    fn new(packages_root_dir: &Path, spec: &PackageSpec) -> Result<Self, PackageError> {
-        // Directory that is 1 level above the package directory.
-        let parent_dir_path = format!("{}/{}", spec.namespace, spec.name);
-        let parent_dir = packages_root_dir.join(&parent_dir_path);
-        let lock_file_name = format!(".{}.lock", spec.version);
-        let lock_file_path = parent_dir.join(&lock_file_name);
-        if !parent_dir.exists() {
-            fs::create_dir_all(parent_dir).map_err(other_err(
-                "failed to create parent directories for lock file",
-            ))?;
-        }
-        Ok(Self {
-            file: fs::OpenOptions::new()
-                .write(true)
-                .create(true)
-                .truncate(true)
-                .open(&lock_file_path)
-                .map_err(other_err("failed to open lock file"))?,
-            path: lock_file_path,
-            tmp_path: format!("{parent_dir_path}/{lock_file_name}"),
-        })
-    }
-
-    /// Try acquiring exclusive lock.
-    fn try_lock(&self) -> io::Result<()> {
-        self.file.try_lock_exclusive()
-    }
-
-    /// Acquire exclusive lock.
-    fn lock(&self) -> Result<(), PackageError> {
-        self.file
-            .lock_exclusive()
-            .map_err(other_err("failed to aquire lock file"))
-    }
-
-    /// Try acquiring the exclusive lock, if failed print an error an wait
-    /// for the lock with blocking.
-    ///
-    /// Returns `Ok(true)` if acquired first try, otherwise `Ok(false)`.
-    fn try_and_lock(&self) -> Result<bool, PackageError> {
-        match self.try_lock() {
-            Ok(_) => {
-                eprintln!("[tmp] acquired the lock file {:?}", self.tmp_path);
-                Ok(true)
-            }
-            Err(ref err) if err.kind() == io::ErrorKind::WouldBlock => {
-                eprintln!("[tmp] couldn't aquire the lock file {:?}", self.tmp_path);
-                eprintln!(
-                    "Waiting for another instance to finish installing the package..."
-                );
-                self.lock()?;
-
-                // // If other instance successfully installed a package after
-                // // waiting for the lock, then there is nothing left to do.
-                // if dir.exists() {
-                //     eprintln!("[tmp] dropped the lock file {:?}", self.tmp_path);
-                //     return Ok(dir);
-                // }
-
-                // eprintln!(
-                //     "Another instance failed to install the package, trying it again."
-                // );
-                Ok(false)
-                // Additional cool/user-friendly logs:
-                // let _write_lock = lock_file.write().unwrap();
-                // println!("checking if package already exists");
-                // if dir.exists() {
-                //     println!("it does! no need to download and unpack");
-                //     return Ok(dir);
-                // } else {
-                //     println!("waited to aquire the file lock and the other instance wasn't able to download it :/");
-                //     todo!()
-                // }
-            }
-            Err(err) => Err(other_err("failed to aquire lock file")(err)),
-        }
-    }
-
-    /// Remove (delete) lock file (if wasn't already by other process).
-    ///
-    /// Will be automatically called when struct is being dropped ([Drop]).
-    fn remove(&self) -> Result<(), PackageError> {
-        if self.path.exists() {
-            fs::remove_file(&self.path)
-                .map_err(other_err("failed to remove lock file"))?;
-        }
-        eprintln!("[tmp] dropped the lock file {:?}", self.tmp_path);
-        Ok(())
-    }
-}
-
-impl Drop for LockFile {
-    fn drop(&mut self) {
-        self.remove().expect("An error occured when dropping a lock file")
-    }
-}
+use crate::lockfile::LockFile;
 
 /// The default Typst registry.
 pub const DEFAULT_REGISTRY: &str = "https://packages.typst.org";
@@ -138,12 +32,6 @@ pub struct PackageStorage {
     downloader: Downloader,
     /// The cached index of the preview namespace.
     index: OnceCell<Vec<PackageInfo>>,
-}
-
-/// Returns mapped `io::Error` to `PackageError::Other` with a custom message.
-/// Intended to be used inside `io::Result`'s `.map_err()` method.
-fn other_err(message: &'static str) -> impl Fn(io::Error) -> PackageError {
-    move |error| PackageError::Other(Some(eco_format!("{message}: {error}")))
 }
 
 impl PackageStorage {
@@ -202,8 +90,8 @@ impl PackageStorage {
                 return Ok(package_dir);
             }
 
-            let lock_file = LockFile::new(cache_dir, spec)?;
-            if !lock_file.try_and_lock()? {
+            let lock_file = LockFile::create_and_lock(cache_dir, spec)?;
+            if lock_file.did_wait() {
                 // If other instance successfully installed a package after
                 // waiting for the lock, then there is nothing left to do.
                 if package_dir.exists() {
